Refactor: Consolidate and remove unused type definitions
This commit addresses the presence of multiple unused, obsolete, and duplicate type definitions within the codebase.

Key changes include:

- Replaced the `HrmData` type alias with `HrmStreamData` to use the core type directly for server-side data.
- Replaced the `SpotifyData` type alias with `SpotifyPlaybackState` for consistency and clarity.
- Introduced a new `ClientHrmData` type that extends `HrmStreamData` with client-specific properties, clarifying the distinction between server-side and client-side data structures.
- Resolved a circular dependency between `WebSocketContext.tsx` and `webSocketReducer.ts` by moving `ClientHrmData` to a central `types/context.ts` file.
- Updated all affected files, including components, services, tests, documentation, and ESLint configuration, to use the new, consolidated types.
- Fixed all linting, build, and test failures that arose during the refactoring process to ensure a stable state.

These changes improve code readability, reduce complexity, and enhance the overall developer experience by simplifying and clarifying the type system.
